updated for test_05
