+++ aisdb_lib/src/decode.rs
@@ -124,11 +124,11 @@ pub fn skipmsg(msg: &str, epoch: &i32) -> Option<(String, i32)> {
     if cols.len() < 6 {
         return Some((msg.to_string(), *epoch));
     }
-    #[cfg(debug_assertions)]
-    println!(
-        "{:?} {:?} {:?} {:?} {:?} {:?}",
-        cols[0], cols[1], cols[2], cols[3], cols[4], cols[5]
-    );
+//     #[cfg(debug_assertions)]
+//     println!(
+//         "{:?} {:?} {:?} {:?} {:?} {:?}",
+//         cols[0], cols[1], cols[2], cols[3], cols[4], cols[5]
+//     );
     let count = str::parse::<u8>(cols[1]).unwrap_or(1);
     let fragment_no = str::parse::<u8>(cols[2]).unwrap_or(1);
     match (cols[0], count, fragment_no, cols[3], cols[4], cols[5]) {
@@ -152,7 +152,7 @@ pub fn filter_vesseldata(
     parser: &mut NmeaParser,
 ) -> Option<(ParsedMessage, i32, bool)> {
     #[cfg(debug_assertions)]
-    println!("{:?} {:?}", epoch, sentence);
+//     println!("{:?} {:?}", epoch, sentence);
 
     match parser.parse_sentence(sentence).ok()? {
         ParsedMessage::VesselDynamicData(vdd) => {
